+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Categories loading
+- Default Seller
+
 ## [3.0.1] - 2020-08-17
 
 ### Changed